don't prematurely close transcript fd
I was closing the transcript fd at the end of run(), because I assumed we'd
wait on a promise to fire before getting that far. But we got past that
point, and Node kept running because of the open listening socket. The
transcript file was closed by then.

For some reason, writes were still successfully going into the transcript
file, but the fd integer itself got re-used by incoming socket connections.

For some weirder reason, at that point transcript lines were also being
written to those sockets (all of them). No idea how it got that broken, but
not closing the fd at the end of run() fixes it.