change guest code layout
Now we execute the 'default' export as an initializer function, passing in an
'argv' object (which will eventually be populated with endowments and/or
config settings from command line arguments and files in the basedir). We
publish whatever it returns as our initial sturdyref. This initializer can
return a Promise/Vow, in which case we wait for it to resolve before
executing the input transcript.

We don't pay attention to any other exports from the guest code. We might
change that in the future, e.g. a resource manager or space-bank underflow
handler, or static metadata about the guest code itself.