argv: add filename:, take ipaddr not hostname
The BASEDIR/argv.json format now knows how to read the contents of files
named by a 'filename' type.

libp2p, at least the way we're using it now, doesn't resolve hostnames, so we
need to record an IP address instead. The 'vat create' argument was renamed
to 'ipaddr' and it now does a quick check for numbers+dots.