add notifyUponResolution
When we send an UnresolvedVow to another Vat, we must remember to tell them
when it resolves. We do this at most once per target vat (so we don't resolve
a single Vow twice), and we must register this notification even if we've
serialized the Vow before.

When we receive an opSend and deliver it (creating a Vow for the result, that
might or might not be resolved yet), we register the new Vow in our
serialization tables so it can be sent over the wire correctly The swissnum
for that Vow was effectively selected by the sender (via swissbase), since
they synthesized a RemoteVow ahead of time (for pipelining). Registering that
Vow also sets up the notification that fires when it resolves.

When we use registerTarget() on the initial occupant of the Vat, we don't
need to notify anybody (since we haven't sent this Vow to anyone yet). In
that case, targetVatID===null and the registration code is skipped.