fix notifyUponResolution
The failing situation was:

* send an unresolvedVow to vat1, causing it to be added to the notification
  table
* resolve the vow, sending an opResolve to vat1
* send the vow to vat2

In this case, vat2 would never be notified of the resolution, because the
'then()' clause had already fired, and no new one was added.

A deeper question is: should we serialize the newly-resolved vow differently
than we did the first time we sent it? We could save a message (but not a
roundtrip) by sending it as a resolvedVow, even though the serialization
cache remembers it as an unresolvedVow.