main.js: manage independent vat-secret
We previously derived this from the host-specific private key (in
BASEDIR/private-id), but that is not the same across all member hosts of a
Quorum vat, which means they create different swissbases, so their messages
diverge. Now we create a random string at vat creation time, and copy it when
converting solo vats into quorum vat hosts.